// src/types/EnumTypes.ts
export enum LoginType {
  AUTHORIZATION = "AUTHORIZATION",
  LOADING = "LOADING",
  SUCCESS = "SUCCESS",
  FAIL = "FAIL",
}

export enum InputBoxType {
  INPUT = "input",
  SELECT = "select",
  TEXTAREA = "textarea",
  DATE_PICKER = "datePicker",
  MULTIPLE_SELECT = "multipleSelect",
}

export enum BooleanType {
  SUCCESS = "SUCCESS",
  FAIL = "FAIL",
  INI = "INI"
}